+++ src/Libraries/Game.Server/Extensions/PaketExtensions.cs
@@ -11,7 +11,7 @@ public static class PaketExtensions
         {
             Id = player.Id,
             Name = player.Name,
-            Class = (byte)player.PlayerClass,
+            Class = player.PlayerClass,
             Level = player.Level,
             Playtime = (uint) TimeSpan.FromMilliseconds(player.PlayTime).TotalMinutes,
             St = player.St,
